Add files via upload
# 3D Skeleton Prediction Quantification using BlazePose

This repository contains a Jupyter notebook that calculates the quantification of the 3D skeleton prediction using BlazePose. The quantification is visualized by representing the loss value as circles on the skeleton. Additionally, the notebook also helps determine the best capture viewpoint of the camera by estimating the radius (r), azimuth, and elevation plots.